Add custom pager script
This is for dealing with less. Less has an option to close
automatically if the content is less than one screenfull. This is great
unless you don't realize that, hit CTRL-D to go down and exit your
terminal. Alas, disabling that in less leaves us with the issue that
git pipes the diff through less even if it's empty. Enter this script.
Exits if empty, pipes through less otherwise