Fuck you google chrome
This fixes a bug/feature of Google Chrome. In a recent (silent) update,
Chrome now brings itself to the foreground whenever opening a new tab,
even when the `-g` argument is passed to `open`. This script now checks
the foreground app before and after, if they differ, open the original.

If only Safari was usable!