Switch to kana's vim indent object
This fixes a bug where if you had something like:

    if foo;
      bar
    fi

    if baz;
      qux
    fi

And, on the top line you did a `gsii` (using the sort motion plugin) it
would ignore the newline between the 2 if statements, and sort the
entire example. Kana's does not have this bug, so only the top if
statement would be sorted.