Don't even get me started
This is the most insane pile of crap. So vim ships with a default
runtime file in `ftplugin.vim`. This file does some interesting stuff
related to filetype plugins. For one, on every `FileType` `autocmd`
fire, it unloads and reloads the current syntax file if possible. This
means if the syntax file follows the standard (but not particularly
documented) pattern of defining a `b:undo_ftplugin` and `b:did_ftplugin`
(specifically with those name), the plugin will be unloaded and
reloaded. This might sound fine, until you realize 2 things. 1)
`FileType` is fired every time `setf` is called. And `setf` is called
everytime `BufRead` fires, and `BufRead` fires a lot. 2) this means if
the syntax plugin sets anything you chose to change in your vimrc, your
setting gets overwritten by the ftplugin. You'd think this wouldn't
happen very often, unless you're like me and despise the `o`
`formatoption` setting, which a lot of ftplugins seem to set for no
reason.

Now to this fix. By adding `FileType` to the list of autocmds I use to
attempt to destroy `o` with, my `FileType` autocmd fires *after* the
unload and reload of the ftplugin, meaning it fixes this crazy crap.

I think in an ideal world, this setting would not be imposed by
ftplugins. It isn't required for a plugin to work as expected _and_ it's
part of the default `formatoptions`