Add custom gx to vimrc
gx is an awesome command defined by netrw. Unfortunately shit hit the
fan when combining opening the URL under the cursor, and fugitive.

Mainly the problem is this. Netrw tries very hard to be smart about what
the `expand("<cfile>")` is. Whether it's a URL, or a remote file URL
etc. Unfortunately to do this it uses a simple `netrw#CheckIfRemote`
function. Which is currently defined as (abbreviated):

```
fun! netrw#CheckIfRemote()
  return expand("%") =~ '^\a\{3,}://'
endfun
```

This would work fine for files that you were accessing over some other
protocol. But unfortunately, fugitive uses `fugitive://path/to/file`
when you're opening git information (say with :Gblame, and pressing
enter on a commit). Because of this, if you use the default gx command
in a commit message in one of these files, it attempts to download it
using curl instead of opening it in your browser.

tl;dr, welp.