Fix bracketed paste with urls and whitespace
This fixes a bug I introduced when I re-enabled bracketed-paste where
urls were not escaped when pasted.

I also took this chance to fix pasting with trailing whitespace, which
is another side effect of having bracketed paste enabled. Now when you
paste a string with leading or trailing whitespace, that whitespace will
be removed.