Add lldb command for printing keypaths
This lldb helper prints the keypath for an object in either Objective-C
or Swift without having to worry about which debugging context you're
currently in. This uses simple dot syntax for the path. Example:

```
self.foo.bar
0x1234.foo.bar
0x1234.self
```

This script handles memory addresses being passed in the Swift debugger
context by casting them to `NSObject` (since that is required for the
type to have `valueForKeyPath`)