set hlsearch in vim when going to next match
Instead of restoring the search on launch
451833048fbe59c8c0f4493d6858dcaf06296f35 because of the issue described
here c99eec409d6a245a6347f868b87f0d2ccaf4a010, instead we can set
hlsearch to on if you navigate to a search that was persisted across
launches. This means it won't show immediately, but if you use the
searches they will show up. This has to be in this mapping and not in
this function because of |function-search-undo|