Update offlineimap
I've been on the previous sha for a while now and this is a bit risky
but I was experiencing this issue:

Server responded: command: APPEND => connection terminated

Updating seems to have fixed it. The configuration change was required
and it didn't re-download all my mail so it must be using sqlite still.