Make zsh vi insert mode always delete
Apparently it's a feature that if you enter insert mode with some
commands like `A`, you cannot delete text that you haven't inserted in
the same insert mode "session". This is the same default behavior as
vim's `backspace` setting, which defaults to nothing, but everyone
immediately sets to include `start` which is the same as this. The way
to fix this for zsh is to remap the keys to always do what you'd expect.