Move to worktree for git PR scripts
This way we avoid cherry pick dances. Creating a worktree takes a bit
for large repos so we use one at a reproducible location. I'm not sure
if it can get borked in /tmp/ after restarts where it's not entirely
cleared. Or maybe when macOS is trying to reclaim disk space.