Fix github delete issue
There used to be an issue described in
2071bd80f5d2b44189a1ed6d2aaf00085eaf719f where deleting an unread github
email from the index would cause issues with the message-hook firing.

It seems that this was caused by save-message, so I changed those uses
to first tag the messages, which is excluded from the message-hook.

This way I could re-enable this in be30b77edd8c44277f537f1832b27937be8d46fd