Fix trustExitCode git config
This isn't read globally but it is read per tool. Without this if you
fix a conflict, but then want to make the mergetool fail (like by using
`:cq` in vim) git would use only the file not having conflict markers to
decide if it was successful or not (since I also have `prompt`
disabled), but I want it to use vim's exit code instead in this case, if
I want it to succeed I will just `ZZ` and it will.