Add BREW_PREFIX var
The thought here is that with Apple Silicon terminals we will have the
new location in the $PATH, and prefer that one over the one in
/usr/local, and with a force Rosetta terminal, or on an Intel machine,
the new location won't be in the $PATH, and theoretically I'll have to
duplicate shared tools across both brew instances. We could probably use
`brew --prefix` for this but it's much much slower.