Add textconv-zip script
In the case a zip file is invalid in the repo, it would stop commands
like `git log -S` from running. In this case I know just ignore them and
print a warning if their output ends up being shown. This is a bit off a
correctness trade-off but I think it's fine realistically.