[git] Add cp-pr alias
This is similar to co-pr except it adds it merges it into the current
branch instead. I'm using (c)herry (p)ick instead of merge because I
don't want to confuse myself with something merging canonically, since
it's just a local action, and ideally it would be a squash and merge so
the commit could easily be dropped if the change merged. Useful in the
git-pile workflows.