[vim] Move from pathogen to vim pack
End of an era. I've been using pathogen for as long as I've been using
vim, and I haven't updated it in over 7 years without a single problem.
Tpope is a legend. Thankfully vim should be updated enough that pack is
usable everywhere at this point, and it works pretty much the same if
you just autoload everything. Pour one out.