Switch to fzf
In my initial testing fzf is just a lot faster than fzy for larger repos. I was trying to improve this before by caching the file searches (which an issue with untracked files today), but even with that it's just too slow. I prefer fzf's incremental loading contents approach, so I can start typing, even if the results take the same amount of time to get there. I'm still using fzy in some places for now, we'll see how that goes.