[vim] Don't jump to wrong position after pasting
Previously in this case:

```
"foo"
```

Where `foo` is highlighted, if I hit paste, it would work correctly.
But then in this case:

```
foo.bar
```

If bar was highlighted and I hit paste, you would end up with` `fooqux.`
instead of `foo.qux`. This is the only way I found to keep both while
also not copying the previous contents into the default register. The
docs don't say that's intentional but it seems like it is.